compatible for higher torch versions
For higher versions than the one in conda_env.yml (e.g. tested on torch=1.12), it gives an error that indicates that the scale of Normal distribution must be greater than 0.

```
ValueError: Expected parameter scale (Tensor of shape (32,32,1024)) of distribution Normal(loc, scale) to satisfy the constraint GreaterThan(lower_bound=0.0), but found invalid values
```

Therefore, a simple fix is to add a small fraction. Another solution is possible for creating a dummy determinisitic distribution to have the same